Update main.py
This code is a Discord bot with several features. The bot can respond to commands such as `/ping` to check latency, `/clear` to delete messages, `/wp` to assign the "wanttoplay" role to a user, `/nw` to remove that role, `!rep+` and `!rep-` to change the reputation of users, `/get_reputation` to get the reputation of a member, `/set_roles` to assign a role to a user, `/remove_role` to remove a role, `/raffle` to conduct a raffle among members, `/mute` to temporarily mute a member, `/delete` to exclude a member, `/ban` to ban a member, and others. The code also contains the `update_status` function, which updates the bot's status every 10 minutes.